+++ packages/yourloops/test/integration/assert/trends.ts
@@ -39,15 +39,36 @@ const STANDARD_DEVIATION_TOOLTIP = 'SD (Standard Deviation): How far values are 
 const CV_TOOLTIP = 'CV (Coefficient of Variation): The ratio of the standard deviation to the mean glucose. For any period greater than 1 day, we calculate the mean of daily CV.'
 
 export const checkTrendsTidelineContainerTooltips = async () => {
+  // Test tooltips when hovering a cbg slice
   const cbgSlice = await screen.findByTestId('cbg-slice-rectangle-innerQuartiles')
   await userEvent.hover(cbgSlice)
-  const tooltips = await screen.findAllByTestId('tooltip')
-  expect(tooltips).toHaveLength(3)
-  expect(tooltips[0]).toHaveTextContent('11:00 am - 11:30 am')
+  const tooltipsOnCbgSlice = await screen.findAllByTestId('tooltip')
+  expect(tooltipsOnCbgSlice).toHaveLength(3)
   const trendsTooltips = screen.getByTestId('trends-tooltips')
-  expect(within(trendsTooltips).getByText('189')).toBeVisible()
-  expect(within(trendsTooltips).getByText('169')).toBeVisible()
+  expect(trendsTooltips).toHaveTextContent('11:00 am - 11:30 am196177')
+
+  // Test tooltips when hovering a cbg circle (cbg slice must still be hovered)
+  const cbgCircles = await screen.findAllByTestId('trends-cbg-circle')
+  expect(cbgCircles).toHaveLength(4)
+  await userEvent.hover(cbgCircles[0])
+  const tooltipsOnCbgCircle = await screen.findAllByTestId('tooltip')
+  expect(tooltipsOnCbgCircle).toHaveLength(4)
+  expect(tooltipsOnCbgCircle[0]).toHaveTextContent('Saturday, January 18')
+  expect(tooltipsOnCbgCircle[1]).toHaveTextContent('11:00 am - 11:30 am')
+  expect(tooltipsOnCbgCircle[2]).toHaveTextContent('196')
+  expect(tooltipsOnCbgCircle[3]).toHaveTextContent('177')
+
+  // Test tooltips when unhovering cbg circle
+  await userEvent.unhover(cbgCircles[0])
+  const tooltipsOnCbgCircle2 = await screen.findAllByTestId('tooltip')
+  expect(tooltipsOnCbgCircle2).toHaveLength(3)
+  expect(tooltipsOnCbgCircle2[0]).toHaveTextContent('11:00 am - 11:30 am')
+  expect(tooltipsOnCbgCircle2[1]).toHaveTextContent('196')
+  expect(tooltipsOnCbgCircle2[2]).toHaveTextContent('177')
+
+  // Test tooltips when unhovering cbg slice
   await userEvent.unhover(cbgSlice)
+  expect(screen.queryAllByTestId('tooltip')).toHaveLength(0)
 }
 
 export const checkTrendsStatsWidgetsTooltips = async () => {